feat(team): add team service to set member
